requirements.txt
# GEPA + SWE-smith + Mini-SWE-Agent (Docker Mode)

# GEPA - Genetic-Pareto prompt optimization framework
-e /home/eecs/shangyin/gepa-optimize-anything

# SWE-smith - Dataset toolkit with Docker container support
# Install: pip install swesmith
# Or from source: git clone https://github.com/SWE-bench/SWE-smith.git && cd SWE-smith && pip install -e .
swesmith
swebench  # Required by swesmith but not declared as dependency

# Mini-SWE-Agent - Lightweight agent for solving GitHub issues
mini-swe-agent

# Docker SDK for Python - Required
docker

# HuggingFace datasets - For loading SWE-smith data
datasets

# LiteLLM - LLM API abstraction
litellm

# Environment variables
python-dotenv

# Utilities
pandas
numpy
tqdm
pytest
